[Domain] Some changes to the domain classes
Add some propeties and relations to Planet and Trade classes.

closes #25

+++ test/unit/com/astrovisor/ClimateSpec.groovy
@@ -15,7 +15,7 @@ class ClimateSpec extends Specification {
     def cleanup() {
     }
 
-    void "test a climate validity"(float aMinTemp, float aMaxTemp, float aMeanTemp, float nbSeasons, boolean expectedState) {
+    void "test a climate validity"(float aMinTemp, float aMaxTemp, float aMeanTemp, int nbSeasons, boolean expectedState) {
         given:"a climate"
         Climate climate = new Climate(aMinTemp, aMaxTemp, aMeanTemp, nbSeasons)
